padronizando para models e excluindo entity #8
