+++ depara/repository.py
@@ -1,6 +1,8 @@
 from typing import List, Optional
+
 from django.db.models import Q
-from .entity import DePara
+
+from .models import DePara
 
 
 class DeparaRepository:
